+++ lib/src/inherited_providers.dart
@@ -2,14 +2,14 @@
 part of 'inherited_object.dart';
 
 class InheritedProviders extends StatelessWidget {
-  final List<InheritedObjectProvider> providers;
+  final List<InheritedObjectProvider> entries;
   final Widget child;
-  const InheritedProviders({super.key, required this.providers, required this.child});
+  const InheritedProviders(this.entries,{super.key, required this.child});
 
   @override
   Widget build(BuildContext context) {
-    return providers.reversed.skip(1).fold(
-      providers.last.copyWithChild(child),
+    return entries.reversed.skip(1).fold(
+      entries.last.copyWithChild(child),
       (previous, current) => current.copyWithChild(previous)
     ); // Chain the providers and return the final widget
   }