+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
-
+## [0.0.4]
+- Changed
+  - `InheritedProviders`: The `providers` parameter is now a positional argument and has been renamed to `entries`.
+- Added
+  - `InheritedObjects`: Enables adding multiple `InheritedObject`s using a single widget.
+  
 ## [0.0.3]
 - Added `InheritedDataProvider` 
   - read-only provider.